Minor updates before testing deployment to a live site
