+++ main/app/contact/page.tsx
@@ -5,9 +5,7 @@ import FormDefault from "@/components/general/FormDefault/FormDefault";
 import contactFormModel from "@/data/models/contact/contact_form.json";
 import parse from "html-react-parser";
 
-interface ContactFormProps {}
-
-const Contact: React.FC<ContactFormProps> = () => {
+const Contact: React.FC = () => {
   return (
     <div className="contact-page container mx-auto px-4 py-16">
       <h1 className="text-4xl font-bold text-gray-800 mb-8">{contactFormModel.title}</h1>
