+++ main/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import NavigationBar from "@/components/NavigationBar";
-import Footer from "@/components/Footer/Footer";
+import NavigationBar from "@/components/general/NavigationBar";
+import Footer from "@/components/general/Footer/Footer";
 import "./globals.css";
 
 const inter = Inter({ subsets: ["latin"] });
