+++ main/types/contact.ts
@@ -1,0 +1,10 @@
+export interface FormDefaultProps {
+  formModel: {
+    form: Array<{
+      name: string;
+      label: string;
+      type: string;
+      required: boolean;
+    }>;
+  };
+}
