--- main/types/projects.ts
@@ -1,4 +1,0 @@
-export interface ProjectCardProps {
-  title?: string;
-  description?: string;
-}
